requirements.txt
astroid==1.6.6
backports.functools-lru-cache==1.5
bcrypt==3.1.7
blinker==1.4
cachetools==3.1.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
colorama==0.4.1
configparser==3.7.4
enum34==1.1.6
facebook-sdk==3.1.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-CLI==0.4.0
Flask-Login==0.4.1
Flask-Moment==0.8.0
Flask-MySQLdb==0.2.0
Flask-Navigation==0.2.0
flask-paginate==0.5.3
Flask-Selfdoc==1.1.0
Flask-Session==0.3.1
Flask-WTF==0.14.2
fudge==1.1.1
gdata==2.0.18
google-api-python-client==1.7.10
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
httplib2==0.13.1
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
mysqlclient==1.4.2.post1
oauthlib==3.0.2
pyasn1==0.4.6
pyasn1-modules==0.2.5
pycparser==2.19
pylint==1.9.4
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
simplejson==3.16.0
singledispatch==3.4.0.3
six==1.12.0
soundcloud==0.5.0
uritemplate==3.0.0
urllib3==1.25.3
Werkzeug==0.15.4
wrapt==1.11.2
WTForms==2.2.1
